% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_clean_strains}
\alias{all_clean_strains}
\title{Data: A list of all post-QC strains used for the analyses}
\format{
A character vector of 1373 lane IDs
\describe{
\item{all_clean_strains}{Character vector of lane IDs}
}
}
\usage{
all_clean_strains
}
\description{
Data: A list of all post-QC strains used for the analyses
}
\keyword{datasets}
