% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_col}
\alias{date_col}
\title{Data: Colourscheme for dates (year)}
\format{
A vector of 25 named characters
\describe{
\item{date_col}{Named character vector of 25 entries}
}
}
\usage{
date_col
}
\description{
Data: Colourscheme for dates (year)
}
\keyword{datasets}
