% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESBL_col}
\alias{ESBL_col}
\title{Data: Colourscheme for the ESBL genes / alleles detected}
\format{
A vector of seven named characters
\describe{
\item{ESBL_col}{Named character vector of seven entries}
}
}
\usage{
ESBL_col
}
\description{
Data: Colourscheme for the ESBL genes / alleles detected
}
\keyword{datasets}
