% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K_col}
\alias{K_col}
\title{Data: Colourscheme for K-types}
\format{
A vector of 27 named characters
\describe{
\item{K_col}{Named character vector of 27 entries}
}
}
\usage{
K_col
}
\description{
Data: Colourscheme for K-types
}
\keyword{datasets}
