% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kp1tree}
\alias{Kp1tree}
\title{Data: K. pneumoniae subsp. pneumoniae mapping tree}
\format{
A "phylo" object ("ape" package)
}
\usage{
Kp1tree
}
\description{
Tree generated as described in manuscript, excluding all STs with only a single isolate in the dataset.
Mapping tree as described recalculated so branch lengths represent SNP distances to use as rPinecone input.
}
\keyword{datasets}
