% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kp2tree}
\alias{Kp2tree}
\title{Data: K. pneumoniae subsp. quasipneumoniae mapping tree}
\format{
A "phylo" object ("ape" package)
}
\usage{
Kp2tree
}
\description{
Tree generated as described in manuscript.
Mapping tree as described recalculated so branch lengths represent SNP distances to use as rPinecone input.
}
\keyword{datasets}
