% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_ST}
\alias{main_ST}
\title{Data: Main sequence types in the dataset}
\format{
A vector of 19 characters
\describe{
\item{main_ST}{Character vector of 19 entries}
}
}
\usage{
main_ST
}
\description{
Data: Main sequence types in the dataset
}
\keyword{datasets}
