% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_ST_col}
\alias{main_ST_col}
\title{Data: Colourscheme for the main ST types}
\format{
A vector of 20 named characters
\describe{
\item{main_ST_col}{Named character vector of 20 entries}
}
}
\usage{
main_ST_col
}
\description{
Data: Colourscheme for the main ST types
}
\keyword{datasets}
