% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{O_colours}
\alias{O_colours}
\title{Data: Colourscheme for the O-Ag loci}
\format{
A vector of 13 named characters
\describe{
\item{O_colours}{Named character vector of 13 entries}
}
}
\usage{
O_colours
}
\description{
Data: Colourscheme for the O-Ag loci
}
\keyword{datasets}
