% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{O_type_colours}
\alias{O_type_colours}
\title{Data: Colourscheme for the O-Ag types as predicted}
\format{
A vector of 17 named characters
\describe{
\item{O_type_colours}{Named character vector of 17 entries}
}
}
\usage{
O_type_colours
}
\description{
Data: Colourscheme for the O-Ag types as predicted
}
\keyword{datasets}
