% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_col}
\alias{sample_col}
\title{Data: Colourscheme for the sample type}
\format{
A vector of 2 named characters
\describe{
\item{sample_col}{Named character vector of 2 entries}
}
}
\usage{
sample_col
}
\description{
Data: Colourscheme for the sample type
}
\keyword{datasets}
