% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_col}
\alias{species_col}
\title{Data: Colourscheme for the (sub)species}
\format{
A vector of 8 named characters
\describe{
\item{species_col}{Named character vector of 8 entries}
}
}
\usage{
species_col
}
\description{
Data: Colourscheme for the (sub)species
}
\keyword{datasets}
