% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST1415tree}
\alias{ST1415tree}
\title{Data: K. pneumoniae subsp. pneumoniae ST1415 tree}
\format{
A "phylo" object ("ape" package)
}
\usage{
ST1415tree
}
\description{
Tree generated as described in manuscript, using only ST14 and ST15 isolates.
Mapping tree as described recalculated so branch lengths represent SNP distances to use as rPinecone input.
}
\keyword{datasets}
