% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST39tree}
\alias{ST39tree}
\title{Data: K. pneumoniae subsp. pneumoniae ST39 tree}
\format{
A "phylo" object ("ape" package)
}
\usage{
ST39tree
}
\description{
Tree generated as described in manuscript, using only ST39 isolates.
Mapping tree as described recalculated so branch lengths represent SNP distances to use as rPinecone input.
}
\keyword{datasets}
