% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_tree}
\alias{test_tree}
\title{Data: Core gene tree based on the output from panaroo}
\format{
A "phylo" object ("ape" package)
}
\usage{
test_tree
}
\description{
Tree generated as described in manuscript,
}
\keyword{datasets}
