% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ward_col}
\alias{ward_col}
\title{Data: Colourscheme for the main wards}
\format{
A vector of 16 named characters
\describe{
\item{ward_col}{Named character vector of 16 entries}
}
}
\usage{
ward_col
}
\description{
Data: Colourscheme for the main wards
}
\keyword{datasets}
